Relocate projection helpers to a dedicated file
