+++ src/projection.rs
@@ -108,4 +108,5 @@ impl Ellipsoid {
     }
 }
 
+mod helpers;
 mod lcc;
